// eshop-web/src/main/resources/client/src/app/layout/components/header/components/header-search/header-search-type.enum.ts
export enum HeaderSearchTypeEnum {
    ALL = 'all',
    ARTICLES = 'articles',
    VEHICLES = 'vehicles',
    BARCODE = 'barcode',
    RETURN = 'return',
    BATTERIES = 'batteries',
    PRODUCT_CATEGORY = 'product_category',
    MORE = 'MORE', // option to see more results
    ARTICLES_DESC = 'ARTICLES_DESC',
    VEHICLES_MOTOR = 'VEHICLES_MOTOR',
    FASTSCAN_ARTICLES = 'fastscan-articles'
}